Tweaks, sort locations by distance and stronger address validation
