+++ apps/site/lib/site_web/controllers/fare_controller.ex
@@ -4,12 +4,13 @@ defmodule SiteWeb.FareController do
   """
   use SiteWeb, :controller
   alias Fares.RetailLocations
+  alias GoogleMaps.Geocode
   import SiteWeb.ViewHelpers, only: [cms_static_page_path: 2]
 
   plug(:meta_description)
 
   @options %{
-    geocode_fn: &GoogleMaps.Geocode.geocode/1,
+    geocode_fn: &Geocode.geocode/1,
     nearby_fn: &Fares.RetailLocations.get_nearby/1
   }
 
@@ -17,7 +18,7 @@ defmodule SiteWeb.FareController do
         %Plug.Conn{assigns: %{date_time: date_time}} = conn,
         %{"id" => "retail-sales-locations"} = params
       ) do
-    {position, formatted} = calculate_position(params, @options.geocode_fn)
+    {position, formatted} = Geocode.calculate_position(params, @options.geocode_fn)
 
     conn
     |> assign(:breadcrumbs, [
@@ -38,39 +39,6 @@ defmodule SiteWeb.FareController do
     check_cms_or_404(conn)
   end
 
-  @spec calculate_position(
-          map(),
-          (String.t() -> GoogleMaps.Geocode.Address.t())
-        ) :: {GoogleMaps.Geocode.Address.t(), String.t()}
-  def calculate_position(%{"latitude" => lat_str, "longitude" => lng_str} = params, geocode_fn) do
-    case {Float.parse(lat_str), Float.parse(lng_str)} do
-      {{lat, ""}, {lng, ""}} ->
-        addr = %GoogleMaps.Geocode.Address{
-          latitude: lat,
-          longitude: lng,
-          formatted: lat_str <> "," <> lng_str
-        }
-
-        parse_geocode_response({:ok, [addr]})
-
-      _ ->
-        params
-        |> Map.delete("latitude")
-        |> Map.delete("longitude")
-        |> calculate_position(geocode_fn)
-    end
-  end
-
-  def calculate_position(%{"location" => %{"address" => address}}, geocode_fn) do
-    address
-    |> geocode_fn.()
-    |> parse_geocode_response()
-  end
-
-  def calculate_position(_params, _geocode_fn) do
-    {%{}, ""}
-  end
-
   @spec current_pass(DateTime.t()) :: String.t()
   def current_pass(%{day: day} = date) when day < 15 do
     Timex.format!(date, "{Mfull} {YYYY}")
@@ -81,17 +49,9 @@ defmodule SiteWeb.FareController do
     Timex.format!(next_month, "{Mfull} {YYYY}")
   end
 
-  defp parse_geocode_response({:ok, [location | _]}) do
-    {location, location.formatted}
-  end
-
-  defp parse_geocode_response(_) do
-    {%{}, ""}
-  end
-
   @spec fare_sales_locations(
-          GoogleMaps.Geocode.Address.t(),
-          (GoogleMaps.Geocode.Address.t() -> [{RetailLocations.Location.t(), float}])
+          Geocode.Address.t(),
+          (Geocode.Address.t() -> [{RetailLocations.Location.t(), float}])
         ) :: [{RetailLocations.Location.t(), float}]
   def fare_sales_locations(%{latitude: _lat, longitude: _long} = position, nearby_fn) do
     nearby_fn.(position)
