+++ apps/site/lib/site_web/views/fare_transformation_view.ex
@@ -3,6 +3,4 @@ defmodule SiteWeb.FareTransformationView do
   View for the Fare Transformation section of the website.
   """
   use SiteWeb, :view
-
-  def alert_icon(), do: svg("icon-alerts-triangle.svg")
 end
